+++ LoL item sets/models/CurrentPatch.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LoL_item_sets.models
+{
+	class CurrentPatch
+	{
+		public string version { get; set; }
+		public DateTime generationDate { get; set; }
+		public string err { get; set; }
+	}
+}