+++ LoL item sets/models/News.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LoL_item_sets.models
+{
+	class News
+	{
+		public string text { get; set; }
+	}
+}